EZP-27476: Rendering of app errors
A custom ExceptionListener decorates the one from Symfony.
After an exception has been handled by it, it registers the exception into
the App.

When the app is rendered to HTML, if it is an exception, the exception's Response is
rendered alone, without any other element. For JSON updates, the exception is returned
as the only property update.